+++ docs_src/pluggables/manual_self_registering.py
@@ -20,6 +20,10 @@ class MyExtension(Extension):
         # Do something here like print a log or whatever you need
         logger.success("Started the extension manually")
 
+        # Add the extension to the extensions of Esmerald
+        # And make it accessible
+        self.app.add_extension("my-extension", self)
+
 
 @get("/home")
 async def home(request: Request) -> JSONResponse:
@@ -34,4 +38,6 @@ async def home(request: Request) -> JSONResponse:
 
 
 app = Esmerald(routes=[Gateway(handler=home)])
-app.add_extension("my-extension", MyExtension)
+
+extension = MyExtension(app=app)
+extension.extend()