+++ esmerald/pluggables/base.py
@@ -1,8 +1,10 @@
 from abc import ABC, abstractmethod
+from inspect import isclass
 from typing import TYPE_CHECKING, Any, Iterator, Optional
 
 from typing_extensions import Annotated, Doc
 
+from esmerald.exceptions import ImproperlyConfigured
 from esmerald.protocols.extension import ExtensionProtocol
 
 if TYPE_CHECKING:  # pragma: no cover
@@ -11,9 +13,7 @@ if TYPE_CHECKING:  # pragma: no cover
 
 class Pluggable:
     """
-    The `Pluggable` is used to create an `Esmerald` pluggable from an `Extension`.
-    When Esmerald receives pluggables, it hooks them into the system and allows
-    the access via anywhere in the application.
+    The `Pluggable` is a wrapper around an Extension to initialize it lazily.
 
     Read more about the [Pluggables](https://esmerald.dev/pluggables/) and learn how to use them.
 
@@ -35,10 +35,12 @@ class Pluggable:
 
     class MyExtension(Extension):
         def __init__(
-            self, app: Optional["Esmerald"] = None, config: PluggableConfig = None, **kwargs: "DictAny"
+            self,
+            app: Optional["Esmerald"] = None,
+            config: PluggableConfig = None,
+            **kwargs: "DictAny",
         ):
             super().__init__(app, **kwargs)
-            self.app = app
 
         def extend(self, config: PluggableConfig) -> None:
             logger.success(f"Successfully passed a config {config.name}")
@@ -48,11 +50,11 @@ class Pluggable:
 
     pluggable = Pluggable(MyExtension, config=my_config)
 
-    app = Esmerald(routes=[], pluggables={"my-extension": pluggable})
+    app = Esmerald(routes=[], extensions={"my-extension": pluggable})
     ```
     """
 
-    def __init__(self, cls: "Extension", **options: Any):
+    def __init__(self, cls: type["ExtensionProtocol"], **options: Any):
         self.cls = cls
         self.options = options
 
@@ -67,35 +69,7 @@ class Pluggable:
         return f"{name}({args})"
 
 
-class BaseExtension(ABC, ExtensionProtocol):  # pragma: no cover
-    """
-    The base for any Esmerald plugglable.
-    """
-
-    def __init__(
-        self,
-        app: Annotated[
-            Optional["Esmerald"],
-            Doc(
-                """
-                An `Esmerald` application instance or subclasses of Esmerald.
-                """
-            ),
-        ] = None,
-        **kwargs: Annotated[
-            Any,
-            Doc("""Any additional kwargs needed."""),
-        ],
-    ):
-        super().__init__(app, **kwargs)
-        self.app = app
-
-    @abstractmethod
-    def extend(self, **kwargs: "Any") -> None:
-        raise NotImplementedError("plug must be implemented by the subclasses.")
-
-
-class Extension(BaseExtension):
+class Extension(ABC, ExtensionProtocol):
     """
     `Extension` object is the one being used to add the logic that will originate
     the `pluggable` in the application.
@@ -115,8 +89,7 @@ class Extension(BaseExtension):
 
     class MyExtension(Extension):
         def __init__(self, app: Optional["Esmerald"] = None, **kwargs: "DictAny"):
-            super().__init__(app, **kwargs)
-            self.app = app
+            super().__init__(app)
             self.kwargs = kwargs
 
         def extend(self, **kwargs: "DictAny") -> None:
@@ -127,3 +100,96 @@ class Extension(BaseExtension):
             # Do something here
     ```
     """
+
+    def __init__(
+        self,
+        app: Annotated[
+            Optional["Esmerald"],
+            Doc(
+                """
+                An `Esmerald` application instance or subclasses of Esmerald.
+                """
+            ),
+        ] = None,
+        **kwargs: Annotated[
+            Any,
+            Doc("""Any additional kwargs needed."""),
+        ],
+    ):
+        super().__init__()
+        self.app = app
+
+    @abstractmethod
+    def extend(self, **kwargs: Any) -> None:
+        raise NotImplementedError("Extension must be implemented by the subclasses.")
+
+
+BaseExtension = Extension
+
+
+class ExtensionDict(dict[str, Extension]):
+    def __init__(self, data: Any = None, *, app: "Esmerald"):
+        super().__init__(data)
+        self.app = app
+        self.delayed_extend: Optional[dict[str, dict[str, Any]]] = {}
+        for k, v in self.items():
+            self[k] = v
+
+    def extend(self) -> None:
+        while self.delayed_extend:
+            key, val = self.delayed_extend.popitem()
+            self[key].extend(**val)
+        self.delayed_extend = None
+
+    def ensure_extension(self, name: str) -> None:
+        """
+        For reordering extension initialization
+
+        class MyExtension2(Extension):
+
+            def extend(self, **kwargs: "DictAny") -> None:
+                '''
+                Function that should always be implemented when extending
+                the Extension class or a `NotImplementedError` is raised.
+                '''
+                self.app.extensions.ensure_extension("foo")
+                # Do something here
+        ```
+
+        """
+        if name not in self:
+            raise ValueError(f"Extension does not exist: {name}")
+        delayed = self.delayed_extend
+        if delayed is None:
+            return
+        val = delayed.pop(name, None)
+        if val is not None:
+            self[name].extend(**val)
+
+    def __setitem__(self, name: Any, value: Any) -> None:
+        if not isinstance(name, str):
+            raise ImproperlyConfigured("Extension names should be in string format.")
+        elif isinstance(value, Pluggable):
+            cls, options = value
+            value = cls(app=self.app, **options)
+            if self.delayed_extend is None:
+                value.extend(**options)
+            else:
+                self.delayed_extend[name] = options
+        elif not isclass(value) and isinstance(value, ExtensionProtocol):
+            if self.delayed_extend is not None:
+                raise ImproperlyConfigured(
+                    "Cannot pass an initialized extension in extensions parameter."
+                )
+        elif isclass(value) and issubclass(value, ExtensionProtocol):
+            value = value(app=self.app)
+            if self.delayed_extend is None:
+                value.extend()
+            else:
+                self.delayed_extend[name] = {}
+        else:
+            raise ImproperlyConfigured(
+                "An extension must subclass from Extension, implement the ExtensionProtocol "
+                "as instance or being wrapped in a Pluggable."
+            )
+        super().__setitem__(name, value)